single truth for version, version argv processing, new project structure with cse and scripts folder, changed vagrant provisioning step, added requirements.txt, renamings
