+++ MANIFEST.in
@@ -1,0 +1,2 @@
+include LICENSE
+include cse/VERSION
