+++ cse/__init__.py
@@ -2,5 +2,14 @@
 CSE - a web crawling and web searching application for news paper
 comments written in Python
 """
+import os
 
-# maybe some globals or imports?
+with open(os.path.join(os.path.dirname(__file__), 'VERSION')) as version_file:
+    version = version_file.read().strip()
+
+__title__ = 'CommentSearchEngine'
+__version__ = version
+__author__ = 'Benedikt Bock, Sebastian Schmidl'
+__license__ = 'MIT'
+
+from .CommentSpider import CommentSpider
