+++ requirements.txt
@@ -1,0 +1,3 @@
+--index-url https://pypi.python.org/simple/
+
+-e .
